% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{score_all}
\alias{score_all}
\title{Score all questionnaires in an export}
\usage{
score_all(obj)
}
\arguments{
\item{obj}{A \code{tallier_export} or \code{tallier_study} object.}
}
\value{
The same object with scores updated in-place.
}
\description{
Convenience wrapper that rescores every result entry in a
\code{tallier_export} or \code{tallier_study} object. This is called automatically
when \code{rescore = TRUE} in \code{\link[=read_scoreme]{read_scoreme()}}.
}
